# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Claudia Vanegas POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [ropa, zapatos, accesorios]
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CUSTOMERS ───────────────────────────────────────────────────────────────
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchaseOrders]
      summary: List all purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, partial, received, cancelled]
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchaseOrders]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchaseOrders]
      summary: Get a purchase order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "404":
          description: Not found
    put:
      operationId: updatePurchaseOrder
      tags: [purchaseOrders]
      summary: Update a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchaseOrders]
      summary: Receive merchandise (partial or full) for a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveOrderInput"
      responses:
        "200":
          description: Updated purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # ─── ACCOUNTS PAYABLE ────────────────────────────────────────────────────────
...
  /reports/sales-by-month:
    get:
      operationId: getReportSalesByMonth
      tags: [reports]
      summary: Sales totals grouped by month
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales by month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesByMonthPoint"
...
  /reports/top-products:
    get:
      operationId: getReportTopProducts
      tags: [reports]
      summary: Top selling products overall
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, code, name, costPrice, salePrice, stock, category, images, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        costPrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        category:
          type: string
          enum: [ropa, zapatos, accesorios]
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [code, name, costPrice, salePrice, stock, category]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        costPrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        category:
          type: string
          enum: [ropa, zapatos, accesorios]
        images:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        costPrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        category:
          type: string
          enum: [ropa, zapatos, accesorios]
        images:
          type: array
          items:
            type: string

    # ─── CUSTOMER ────────────────────────────────────────────────────────────
...
    PurchaseOrderItem:
      type: object
      required: [id, productId, productName, qtyOrdered, qtyReceived, unitCost]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        qtyOrdered:
          type: integer
        qtyReceived:
          type: integer
        unitCost:
          type: number
    PurchaseOrderItemInput:
      type: object
      required: [productId, qtyOrdered, unitCost]
      properties:
        productId:
          type: integer
        qtyOrdered:
          type: integer
        unitCost:
          type: number
    PurchaseOrder:
      type: object
      required: [id, supplierId, supplierName, guideNumber, paymentType, status, total, items, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        guideNumber:
          type: string
        paymentType:
          type: string
          enum: [contado, credito]
        status:
          type: string
          enum: [pending, partial, received, cancelled]
        total:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        createdAt:
          type: string
          format: date-time
    PurchaseOrderInput:
      type: object
      required: [supplierId, guideNumber, paymentType, items]
      properties:
        supplierId:
          type: integer
        guideNumber:
          type: string
        paymentType:
          type: string
          enum: [contado, credito]
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
    PurchaseOrderUpdate:
      type: object
      properties:
        guideNumber:
          type: string
        paymentType:
          type: string
          enum: [contado, credito]
        status:
          type: string
          enum: [pending, partial, received, cancelled]
        notes:
          type: string
    ReceiveItemInput:
      type: object
      required: [purchaseOrderItemId, qtyReceived]
      properties:
        purchaseOrderItemId:
          type: integer
        qtyReceived:
          type: integer
    ReceiveOrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReceiveItemInput"

    # ─── ACCOUNTS PAYABLE ────────────────────────────────────────────────────
...
    TopProduct:
      type: object
      required: [productId, productName, category, totalQty, totalRevenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        category:
          type: string
        totalQty:
          type: integer
        totalRevenue:
          type: number

    # ─── REPORTS ─────────────────────────────────────────────────────────────
    SalesByMonthPoint:
      type: object
      required: [month, total, count]
      properties:
        month:
          type: string
        total:
          type: number
        count:
          type: integer
...