# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Claudia Vanegas POS API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCategory } from './productInputCategory';

export interface ProductInput {
  code: string;
  name: string;
  description?: string;
  costPrice: number;
  salePrice: number;
  stock: number;
  category: ProductInputCategory;
  images?: string[];
}